new structure for the project. all the routes now live inside a package with the same name, and the database is now initialized inside a directory called 'database', this makes it reachable from any point of the full app
